% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\name{CenterKernel}
\alias{CenterKernel}
\title{Center a kernel matrix}
\usage{
CenterKernel(K)
}
\arguments{
\item{K}{A n*n symmetric matrix}
}
\value{
A centered matrix
}
\description{
Center a kernel matrix
}
