% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_parameter_objects.R
\name{combine_parameter_objects}
\alias{combine_parameter_objects}
\title{Get clean names of model paramters}
\usage{
combine_parameter_objects(x, ..., split_by = c("Effects", "Component",
  "Response"))
}
\arguments{
\item{x}{A fitted model.}

\item{...}{One or more objects (data frames), which contain information about
the model parameters and related statistics (like confidence intervals, HDI,
ROPE, ...).}

\item{split_by}{If \code{NULL}, the merged data frame is returned. Else, the
data frame is split into a list, split by the values from those columns defined
in \code{split_by}. \code{split_by} should be a character vector with one or
more of the following elements: \code{"Effects"}, \code{"Component"},
\code{"Response"} and \code{"Group"}.}
}
\value{
A data frame or a list.
}
\description{
This function
}
\examples{
library(bayestestR)
model <- download_model("brms_zi_2")
x <- hdi(model, effects = "all", component = "all")
x
combine_parameter_objects(model, x)

}
