% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compact.R
\name{compact_character}
\alias{compact_character}
\title{Remove empty strings from character}
\usage{
compact_character(x)
}
\arguments{
\item{x}{A single character or a vector of characters.}
}
\value{
A character or a character vector with empty strings removed.
}
\description{
Remove empty strings from character
}
\examples{
compact_character(c("x", "y", NA))
compact_character(c("x", "NULL", "", "y"))

}
