% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\name{ComputeBandwidth}
\alias{ComputeBandwidth}
\title{Compute the bandwidth hyper-parameters for a INSiGHT object}
\usage{
ComputeBandwidth(
  object,
  k.NN = 100,
  kernel.option = c("single", "mixture"),
  quantile.bandwidth = 0.1
)
}
\arguments{
\item{object}{INSiGHT object}

\item{k.NN}{The maximum number of nearest neighbours to compute for the kd-tree.}

\item{kernel.option}{Charater vector specifying whether a "single" kernel
or a "mixture" of multiple kernels will be contructed for spatial similarity.}

\item{quantile.bandwidth}{(default 0.1) A number of probability with values in \eqn{[0,1]},
for computing the quantile of pair-wise distances.
Works only when \code{kernel.option = "single"}.
When \code{kernel.option = "mixture"}, five quantiles (0.1, 0.3, 0.5, 0.7, 0.9) will computed.}
}
\value{
A INSiGHT object.
}
\description{
Conpute the bandwidth hyperparameters of the Gaussian kernels for a INSiGHT,
using the quantiles of pair-wise distances of spots.
Only the pair-wise distances of the k-nearest neighbors for each spot will be calculated,
using a kd-tree.
}
\keyword{internal}
