% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\name{createINSiGHTobject}
\alias{createINSiGHTobject}
\title{Create the INSiGHT object.}
\usage{
createINSiGHTobject(
  data.input,
  spatial.locs,
  sample.names = NULL,
  covariates = NULL,
  weights.list = NULL
)
}
\arguments{
\item{data.input}{A list containing original gene expression matrices
of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each gene expression matrix is of dimension \eqn{G \times N}, where \eqn{G} rows represent \eqn{G} genes and
\eqn{N} columns represent \eqn{N} spots across this sample.}

\item{spatial.locs}{A list containing spatial coordinate matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each spatial coordinate matrix is of dimension \eqn{N \times 2},
where \eqn{N} rows represent \eqn{N} spots across this sample,
and 2 columns represents x axis and y axis.}

\item{sample.names}{A character string vector specifying the sample IDs of all the samples to analyze.
The vector is of length \eqn{M}, where \eqn{M} is the number of samples.}

\item{covariates}{(default NULL) A list containing covariate matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each covariate matrix is of dimension \eqn{N \times p},
where \eqn{N} rows represent \eqn{N} spots across this sample,
and \eqn{p} columns represents \eqn{p} covariates.}

\item{weights.list}{A list containing weight matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each covariate matrix is of dimension \eqn{N \times N},
where \eqn{N} is the number of spots across this sample.}
}
\value{
A INSiGHT object.
}
\description{
Create the INSiGHT object.
}
