% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.badlink}
\alias{.badlink}
\title{helper-function, telling user if family / distribution is supported or not}
\usage{
.badlink(link, family, verbose = TRUE)
}
\description{
helper-function, telling user if family / distribution is supported or not
}
\keyword{internal}
