% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.between_subject_variance}
\alias{.between_subject_variance}
\title{random intercept-variances, i.e.
between-subject-variance (tau 00)}
\usage{
.between_subject_variance(vals, x)
}
\description{
random intercept-variances, i.e.
between-subject-variance (tau 00)
}
\keyword{internal}
