% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.collapse_cond}
\alias{.collapse_cond}
\title{glmmTMB returns a list of model information, one for conditional
and one for zero-inflated part, so here we "unlist" it, returning
only the conditional part.}
\usage{
.collapse_cond(x)
}
\description{
glmmTMB returns a list of model information, one for conditional
and one for zero-inflated part, so here we "unlist" it, returning
only the conditional part.
}
\keyword{internal}
