% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.compute_variance_dispersion}
\alias{.compute_variance_dispersion}
\title{Get dispersion-specific variance}
\usage{
.compute_variance_dispersion(x, vals, faminfo, obs.terms)
}
\description{
Get dispersion-specific variance
}
\keyword{internal}
