% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.compute_variance_distribution}
\alias{.compute_variance_distribution}
\title{Calculate Distribution-specific variance (Nakagawa et al. 2017)}
\usage{
.compute_variance_distribution(x, var.cor, faminfo, name, verbose = TRUE)
}
\description{
Calculate Distribution-specific variance (Nakagawa et al. 2017)
}
\keyword{internal}
