% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.compute_variance_random}
\alias{.compute_variance_random}
\title{Compute variance associated with a random-effects term (Johnson 2014)}
\usage{
.compute_variance_random(terms, x, vals)
}
\description{
Compute variance associated with a random-effects term (Johnson 2014)
}
\keyword{internal}
