% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.get_variance_default}
\alias{.get_variance_default}
\title{Get distribution-specific variance for general and
undefined families / link-functions}
\usage{
.get_variance_default(x, mu, verbose)
}
\description{
Get distribution-specific variance for general and
undefined families / link-functions
}
\keyword{internal}
