% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.get_variance_information}
\alias{.get_variance_information}
\title{store essential information on coefficients, model matrix and so on
as list, since we need these information throughout the functions to
calculate the variance components...}
\usage{
.get_variance_information(x, faminfo, name_fun = "get_variances",
  verbose = TRUE)
}
\description{
store essential information on coefficients, model matrix and so on
as list, since we need these information throughout the functions to
calculate the variance components...
}
\keyword{internal}
