% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.get_variance_nbinom_family}
\alias{.get_variance_nbinom_family}
\title{Get distributional variance for nbinom-family}
\usage{
.get_variance_nbinom_family(x, null_model, mu, sig, faminfo)
}
\description{
Get distributional variance for nbinom-family
}
\keyword{internal}
