% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.get_variance_tweedie_family}
\alias{.get_variance_tweedie_family}
\title{Get distributional variance for tweedie-family}
\usage{
.get_variance_tweedie_family(x, mu, phi)
}
\description{
Get distributional variance for tweedie-family
}
\keyword{internal}
