% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.null_model}
\alias{.null_model}
\title{Null model is needed to calculate the mean for the model's response,
which we need to compute the distribution-specific variance
(see .variance_distributional())}
\usage{
.null_model(model, verbose = TRUE)
}
\description{
Null model is needed to calculate the mean for the model's response,
which we need to compute the distribution-specific variance
(see .variance_distributional())
}
\keyword{internal}
