% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.random_slopes}
\alias{.random_slopes}
\title{return names of random slopes}
\usage{
.random_slopes(random.effects = NULL, model = NULL)
}
\description{
return names of random slopes
}
\keyword{internal}
