% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_distributional}
\alias{.variance_distributional}
\title{This is the core-function to calculate the distribution-specific variance
Nakagawa et al. 2017 propose three different methods, here we only rely
on the lognormal-approximation.}
\usage{
.variance_distributional(x, faminfo, sig, name, verbose = TRUE)
}
\description{
This is the core-function to calculate the distribution-specific variance
Nakagawa et al. 2017 propose three different methods, here we only rely
on the lognormal-approximation.
}
\keyword{internal}
