% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_family_beta}
\alias{.variance_family_beta}
\title{Get distributional variance for beta-family}
\usage{
.variance_family_beta(x, mu, phi)
}
\description{
Get distributional variance for beta-family
}
\keyword{internal}
