% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_family_default}
\alias{.variance_family_default}
\title{Get distribution-specific variance for general and
undefined families / link-functions}
\usage{
.variance_family_default(x, mu, verbose)
}
\description{
Get distribution-specific variance for general and
undefined families / link-functions
}
\keyword{internal}
