% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_family_nbinom}
\alias{.variance_family_nbinom}
\title{Get distributional variance for nbinom-family}
\usage{
.variance_family_nbinom(x, mu, sig, faminfo)
}
\description{
Get distributional variance for nbinom-family
}
\keyword{internal}
