% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_family_poisson}
\alias{.variance_family_poisson}
\title{Get distributional variance for poisson-family}
\usage{
.variance_family_poisson(x, mu, faminfo)
}
\description{
Get distributional variance for poisson-family
}
\keyword{internal}
