% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_family_tweedie}
\alias{.variance_family_tweedie}
\title{Get distributional variance for tweedie-family}
\usage{
.variance_family_tweedie(x, mu, phi)
}
\description{
Get distributional variance for tweedie-family
}
\keyword{internal}
