% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variances.R
\name{.variance_zip}
\alias{.variance_zip}
\title{For zero-inflated poisson models, the distributional variance
is based on Zuur et al. 2012}
\usage{
.variance_zip(model, faminfo, family_var)
}
\description{
For zero-inflated poisson models, the distributional variance
is based on Zuur et al. 2012
}
\keyword{internal}
