% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_column_order.R
\name{easystats_columns}
\alias{easystats_columns}
\title{Easystats columns}
\usage{
easystats_columns(select = "all")
}
\arguments{
\item{select}{String, indicating which columns to return.}
}
\value{
A character vector with all (or selected) column names that are
in use across the easystats eco-system.
}
\description{
Returns all valid column names that are used or defined across easystats
packages as character vector.
}
\examples{
easystats_columns("uncertainty")

}
