% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{tidy.parameters_model}
\alias{tidy.parameters_model}
\title{Tidy methods for easystats-objects}
\usage{
tidy.parameters_model(x, ...)
}
\arguments{
\item{x}{A data frame of class \code{parameters_model}, as returned by \code{\link[parameters:model_parameters]{model_parameters()}}.}

\item{...}{Currently not used.}
}
\description{
Tidy methods for easystats-objects. \code{tidy()} usually returns
  the same data frame that was used as input, however, with broom-alike
  column names.
}
\examples{
if (require("parameters") && require("generics")) {
  model <- lm(mpg ~ wt + cyl, data = mtcars)
  mp <- model_parameters(model)

  as.data.frame(mp)
  tidy(mp)
}

}
\seealso{
\code{\link{standardize_names}}
}
