% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{efc_insight}
\alias{efc_insight}
\title{Sample dataset from the EFC Survey}
\description{
A sample data set, internally used in tests. Consists of 28
variables from 908 observations. The data set is part of the EUROFAMCARE
project, a cross-national survey on informal caregiving in Europe. Useful
when testing on "real-life" data sets, including random missing values. This
data set also has value and variable label attributes.
}
\keyword{data}
