% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distributional.R
\name{find_distributional}
\alias{find_distributional}
\title{Find distributional parameters from models}
\usage{
find_distributional(x, ...)
}
\arguments{
\item{x}{A model of class \code{brmsfit}.}

\item{...}{Currently not used.}
}
\value{
All available distributional parameters used in the model.
}
\description{
Returns all distributional parameters from brms-models, like
dispersion, sigma, kappa, phi, or beta...
}
