% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_random.R
\name{find_random}
\alias{find_random}
\alias{find_random.glmmTMB}
\title{Get names of random effect terms}
\usage{
find_random(x, split_nested = FALSE, ...)

\method{find_random}{glmmTMB}(x, split_nested = FALSE,
  component = c("all", "conditional", "zi", "zero_inflated",
  "dispersion"), flatten = FALSE, ...)
}
\arguments{
\item{x}{A fitted mixed model.}

\item{split_nested}{Logical, if \code{TRUE}, terms from nested random
effects will be returned as separeted elements, not as single string
with colon. See 'Examples'.}

\item{...}{Currently not used.}

\item{component}{Should all predictor variables, predictor variables for the
conditional model, the zero-inflated part of the model or the dispersion
model be returned? Only applies to models with zero-inflated and/or
dispersion formula. May be abbreviated.}

\item{flatten}{Logical, if \code{TRUE}, the name of model terms are returned
as a single character, not as list.}
}
\value{
The name(s) of the random effects as character vector; for mixed models
   with zero-inflated component, if \code{component = "all"}, a list with
   the names of the random effects from the conditional and zero-inflated
   model parts.
}
\description{
to do...
}
\examples{
library(lme4)
data(sleepstudy)
sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
sleepstudy$mysubgrp <- NA
for (i in 1:5) {
  filter_group <- sleepstudy$mygrp == i
  sleepstudy$mysubgrp[filter_group] <-
  sample(1:30, size = sum(filter_group), replace = TRUE)
}

m <- lmer(
  Reaction ~ Days + (1 | mygrp / mysubgrp) + (1 | Subject),
  data = sleepstudy
)

find_random(m)
find_random(m, split_nested = TRUE)

}
