% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_restoretype.R
\name{force_numeric}
\alias{force_numeric}
\title{Convert to Numeric if possible}
\usage{
force_numeric(x)
}
\arguments{
\item{x}{A vector to be converted.}
}
\value{
Numeric vector (if possible)
}
\description{
Tries to convert vector to numeric if possible (if no warnings or errors). Otherwise, leaves it as is.
}
\examples{
force_numeric(c("1", "2"))
force_numeric(c("1", "2", "A"))
}
