% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_approach.R
\name{format_approach}
\alias{format_approach}
\title{Guess requested statistical method}
\usage{
format_approach(method)
}
\arguments{
\item{method}{Character string describing the method/type of statistics.}
}
\description{
Relevant for packages (e.g., `ggstatsplot`, `statsExpressions`, etc.), where
different types of statistical methods are supported. This switch function
converts strings entered by users to a common pattern.

Possible outputs are:
\itemize{
 \item{parametric}
 \item{non-parametric}
 \item{robust}
 \item{bayesian}
}
}
\examples{
format_approach("p")
format_approach("bf")
}
