% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_capitalize.R
\name{format_capitalize}
\alias{format_capitalize}
\title{Capitalizes the first letter in a string}
\usage{
format_capitalize(x, verbose = TRUE)
}
\arguments{
\item{x}{A character vector or a factor. The latter is coerced to character.
All other objects are returned unchanged.}

\item{verbose}{Toggle warnings.}
}
\value{
\code{x}, with first letter capitalized.
}
\description{
This function converts the first letter in a string into upper case.
}
\examples{
format_capitalize("hello")
format_capitalize(c("hello", "world"))
unique(format_capitalize(iris$Species))
}
