% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auxiliary.R
\name{get_auxiliary}
\alias{get_auxiliary}
\title{Get auxiliary parameters from models}
\usage{
get_auxiliary(x, type = c("sigma", "dispersion"), verbose = TRUE, ...)
}
\arguments{
\item{x}{A model.}
}
\value{
The requested auxiliary parameter, or \code{NULL} if this information
could not be accessed.
}
\description{
Returns the requested auxiliary parameters from models, like
dispersion, sigma, or beta...
}
\details{
Currently, only sigma and the dispersion parameter are returned, and
only for a limited set of models.
}
\examples{
# from ?glm
clotting <- data.frame(
  u = c(5, 10, 15, 20, 30, 40, 60, 80, 100),
  lot1 = c(118, 58, 42, 35, 27, 25, 21, 19, 18),
  lot2 = c(69, 35, 26, 21, 18, 16, 13, 12, 12)
)
model <- glm(lot1 ~ log(u), data = clotting, family = Gamma())
get_auxiliary(model) # same as summary(model)$dispersion
}
