% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deviance.R
\name{get_deviance}
\alias{get_deviance}
\title{Model Deviance}
\usage{
get_deviance(x, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Not used.}
}
\value{
The model deviance.
}
\description{
Returns model deviance (see \code{stats::deviance()}).
}
\examples{
data(mtcars)
x <- lm(mpg ~ cyl, data = mtcars)
get_deviance(x)

if (require("rstanarm")) {
  x <- rstanarm::stan_glm(mpg ~ cyl, data = mtcars, refresh = 0)
  get_deviance(x)
}
}
