% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modelmatrix.R
\name{get_modelmatrix}
\alias{get_modelmatrix}
\title{Model Matrix}
\usage{
get_modelmatrix(x, ...)
}
\arguments{
\item{x}{An object.}
}
\description{
Creates a design matrix from the description. Any character variables are coerced to factors.
}
\examples{
data(mtcars)

model <- lm(am ~ vs, data = mtcars)
get_modelmatrix(model)

}
