% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_mfx.R
\name{get_parameters.betamfx}
\alias{get_parameters.betamfx}
\alias{get_parameters.logitmfx}
\title{Get model parameters from marginal effects models}
\usage{
\method{get_parameters}{betamfx}(
  x,
  component = c("all", "conditional", "precision", "marginal"),
  ...
)

\method{get_parameters}{logitmfx}(x, component = c("all", "conditional", "marginal"), ...)
}
\arguments{
\item{x}{A fitted model.}

\item{component}{Indicates which model component should be returned. E.g.,
should all predictor variables, predictor variables for the conditional
model, the zero-inflated part of the model, the dispersion term or the
instrumental variables be returned? See section \emph{Model Components} for
further details. May be abbreviated. Note that the \emph{conditional} component is
also called \emph{count} or \emph{mean} component, depending on the model.}

\item{...}{Currently not used.}
}
\value{
A data frame with three columns: the parameter names, the related
point estimates and the component.
}
\description{
Returns the coefficients from a model.
}
\section{Model components}{

Possible values for the \code{component} argument depend on the model class.
Following are valid options:
\itemize{
\item \code{"all"}: returns all model components, applies to all models, but will only
have an effect for models with more than just the conditional model component.
\item \code{"conditional"}: only returns the conditional component, i.e. "fixed effects"
terms from the model. Will only have an effect for models with more than
just the conditional model component.
\item \code{"smooth_terms"}: returns smooth terms, only applies to GAMs (or similar
models that may contain smooth terms).
\item \code{"zero_inflated"} (or \code{"zi"}): returns the zero-inflation component.
\item \code{"dispersion"}: returns the dispersion model component. This is common
for models with zero-inflation or that can model the dispersion parameter.
\item \code{"instruments"}: for instrumental-variable or some fixed effects regression,
returns the instruments.
\item \code{"nonlinear"}: for non-linear models (like models of class \code{nlmerMod} or
\code{nls}), returns staring estimates for the nonlinear parameters.
\item \code{"correlation"}: for models with correlation-component, like \code{gls}, the
variables used to describe the correlation structure are returned.
\item \code{"location"}: returns location parameters such as \code{conditional},
\code{zero_inflated}, \code{smooth_terms}, or \code{instruments} (everything that are
fixed or random effects - depending on the \code{effects} argument - but no
auxiliary parameters).
\item \code{"distributional"} (or \code{"auxiliary"}): components like \code{sigma}, \code{dispersion},
\code{beta} or \code{precision} (and other auxiliary parameters) are returned.
}

\strong{Special models}

Some model classes also allow rather uncommon options. These are:
\itemize{
\item \code{mhurdle}: \code{"infrequent_purchase"}, \code{"ip"}, and \code{"auxiliary"}
\item \code{BGGM}: \code{"correlation"} and \code{"intercept"}
\item \code{BFBayesFactor}, \code{glmx}: \code{"extra"}
\item \code{averaging}:\code{"conditional"} and \code{"full"}
\item \code{mjoint}: \code{"survival"}
}
}

\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_parameters(m)
}
