% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_mixed.R
\name{get_parameters.glmm}
\alias{get_parameters.glmm}
\alias{get_parameters.coxme}
\alias{get_parameters.nlmerMod}
\alias{get_parameters.merMod}
\alias{get_parameters.glmmTMB}
\alias{get_parameters.glimML}
\title{Get model parameters from mixed models}
\usage{
\method{get_parameters}{glmm}(x, effects = c("all", "fixed", "random"), ...)

\method{get_parameters}{coxme}(x, effects = c("fixed", "random"), ...)

\method{get_parameters}{nlmerMod}(
  x,
  effects = c("fixed", "random"),
  component = c("all", "conditional", "nonlinear"),
  ...
)

\method{get_parameters}{merMod}(x, effects = c("fixed", "random"), ...)

\method{get_parameters}{glmmTMB}(
  x,
  effects = c("fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
  ...
)

\method{get_parameters}{glimML}(x, effects = c("fixed", "random", "all"), ...)
}
\arguments{
\item{x}{A fitted model.}

\item{effects}{Should parameters for fixed effects, random effects
or both be returned? Only applies to mixed models. May be abbreviated.}

\item{...}{Currently not used.}

\item{component}{Which type of parameters to return, such as parameters for
the conditional model, the zero-inflated part of the model or the
dispersion term? Applies to models with zero-inflated and/or dispersion
formula. Note that the \emph{conditional} component is also called
\emph{count} or \emph{mean} component, depending on the model. There are
three convenient shortcuts: \code{component = "all"} returns all possible
parameters. If \code{component = "location"}, location parameters such as
\code{conditional} or \code{zero_inflated} are returned (everything that
are fixed or random effects - depending on the \code{effects} argument -
but no auxiliary parameters). For \code{component = "distributional"} (or
\code{"auxiliary"}), components like \code{sigma} or \code{dispersion} (and
other auxiliary parameters) are returned.}
}
\value{
If \code{effects = "fixed"}, a data frame with two columns: the
parameter names and the related point estimates. If \code{effects = "random"}, a list of data frames with the random effects (as returned by
\code{ranef()}), unless the random effects have the same simplified
structure as fixed effects (e.g. for models from \strong{MCMCglmm}).
}
\description{
Returns the coefficients from a model.
}
\details{
In most cases when models either return different "effects" (fixed,
random) or "components" (conditional, zero-inflated, ...), the arguments
\code{effects} and \code{component} can be used.
}
\section{Model components}{

Possible values for the \code{component} argument depend on the model class.
Following are valid options:
\itemize{
\item \code{"all"}: returns all model components, applies to all models, but will only
have an effect for models with more than just the conditional model component.
\item \code{"conditional"}: only returns the conditional component, i.e. "fixed effects"
terms from the model. Will only have an effect for models with more than
just the conditional model component.
\item \code{"smooth_terms"}: returns smooth terms, only applies to GAMs (or similar
models that may contain smooth terms).
\item \code{"zero_inflated"} (or \code{"zi"}): returns the zero-inflation component.
\item \code{"dispersion"}: returns the dispersion model component. This is common
for models with zero-inflation or that can model the dispersion parameter.
\item \code{"instruments"}: for instrumental-variable or some fixed effects regression,
returns the instruments.
\item \code{"nonlinear"}: for non-linear models (like models of class \code{nlmerMod} or
\code{nls}), returns staring estimates for the nonlinear parameters.
\item \code{"correlation"}: for models with correlation-component, like \code{gls}, the
variables used to describe the correlation structure are returned.
\item \code{"location"}: returns location parameters such as \code{conditional},
\code{zero_inflated}, \code{smooth_terms}, or \code{instruments} (everything that are
fixed or random effects - depending on the \code{effects} argument - but no
auxiliary parameters).
\item \code{"distributional"} (or \code{"auxiliary"}): components like \code{sigma}, \code{dispersion},
\code{beta} or \code{precision} (and other auxiliary parameters) are returned.
}

\strong{Special models}

Some model classes also allow rather uncommon options. These are:
\itemize{
\item \code{mhurdle}: \code{"infrequent_purchase"}, \code{"ip"}, and \code{"auxiliary"}
\item \code{BGGM}: \code{"correlation"} and \code{"intercept"}
\item \code{BFBayesFactor}, \code{glmx}: \code{"extra"}
\item \code{averaging}:\code{"conditional"} and \code{"full"}
\item \code{mjoint}: \code{"survival"}
}
}

\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_parameters(m)
}
