% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_htest.R
\name{get_parameters.htest}
\alias{get_parameters.htest}
\title{Get model parameters from htest-objects}
\usage{
\method{get_parameters}{htest}(x, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}
}
\value{
A data frame with two columns: the parameter names and the related
  point estimates.
}
\description{
Returns the parameters from a hypothesis test.
}
\examples{
data(sleep)
get_parameters(t.test(extra ~ 1, data = sleep))
}
