% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted.R
\name{get_predicted.lrm}
\alias{get_predicted.lrm}
\title{the rms::lrm function produces an object of class c("lrm", "rms", glm"). The
\code{get_predicted.glm} function breaks when trying to calculate standard errors,
so we use the default method.}
\usage{
\method{get_predicted}{lrm}(x, data = NULL, verbose = TRUE, ...)
}
\description{
the rms::lrm function produces an object of class c("lrm", "rms", glm"). The
\code{get_predicted.glm} function breaks when trying to calculate standard errors,
so we use the default method.
}
