% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{Get the data from a model's predictor variables}
\usage{
get_predictors(x)
}
\arguments{
\item{x}{A fitted model.}
}
\value{
The data of all predictor variables, as data frame.
}
\description{
Returns the data of all predictor variables (fixed effects).
}
\examples{
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
head(get_predictors(m))
}
