% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\alias{get_residuals.default}
\title{Extract model residuals}
\usage{
get_residuals(x, ...)

\method{get_residuals}{default}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Not used.}

\item{verbose}{Toggle warnings and messages.}
}
\value{
The residuals, or \code{NULL} if this information could not be accessed.
}
\description{
Returns the residuals from regression models.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_residuals(m)
}
