% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulated.R
\name{get_simulated}
\alias{get_simulated}
\alias{get_simulated.lm}
\alias{get_simulated.betareg}
\alias{get_simulated.glmmTMB}
\alias{get_simulated.merMod}
\alias{get_simulated.default}
\alias{get_simulated.data.frame}
\title{Simulate values from fitted models}
\usage{
get_simulated(x, ...)

\method{get_simulated}{lm}(x, data = NULL, iterations = 1, seed = NULL, ...)

\method{get_simulated}{betareg}(x, data = NULL, iterations = 1, seed = NULL, ...)

\method{get_simulated}{glmmTMB}(x, data = NULL, iterations = 1, seed = NULL, re.form = NULL, ...)

\method{get_simulated}{merMod}(
  x,
  data = NULL,
  iterations = 1,
  seed = NULL,
  use.u = FALSE,
  re.form = NA,
  newparams = NULL,
  family = NULL,
  cluster.rand = rnorm,
  allow.new.levels = FALSE,
  na.action = na.pass,
  ...
)

\method{get_simulated}{default}(x, ...)

\method{get_simulated}{data.frame}(x, data = NULL, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Additional arguments (currently unused for \code{lm}).}

\item{data}{An optional data frame in which to evaluate predictions before
simulation. This can be a data grid created with \code{\link[=get_datagrid]{get_datagrid()}}.}

\item{iterations}{Number of response vectors to simulate.}

\item{seed}{An optional integer random seed.}

\item{re.form}{For \code{glmmTMB} models, random effects formula passed to
simulation (\code{NULL}, \code{NA} or \code{~0}).}
}
\value{
A data frame with one column per simulation (\code{iter_1}, \code{iter_2}, ...).
The attribute \code{seed} contains information about the RNG state used.
}
\description{
Simulate responses from fitted statistical models.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl, data = mtcars)

# Simulations on the original data
get_simulated(m, iterations = 2, seed = 123)

# Simulations on a data grid
dg <- get_datagrid(m, wt = c(2, 3), cyl = c(4, 6))
get_simulated(dg, m, iterations = 2, seed = 123)

}
