% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_method.R
\name{get_statistical_method}
\alias{get_statistical_method}
\title{Guess requested statistical method}
\usage{
get_statistical_method(method)
}
\arguments{
\item{method}{Character string describing the method/type of statistics.}
}
\description{
Relevant for packages (e.g., `ggstatsplot`, `statsExpressions`, etc.), where
different types of statistical methods are supported. This switch function
converts strings entered by users to a common pattern.

Possible outputs are:
\itemize{
 \item{parametric}
 \item{non-parametric}
 \item{robust}
 \item{bayesian}
}
}
\examples{
get_statistical_method("p")
get_statistical_method("bf")
}
