% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variances.R
\name{get_variances}
\alias{get_variances}
\alias{get_resid_variance}
\alias{get_fixef_variance}
\alias{get_ranef_variance}
\alias{get_dist_variance}
\alias{get_disp_variance}
\title{Get variance components from random effects models}
\usage{
get_variances(x, component = c("all", "fixef", "ranef", "resid", "dist",
  "disp"), ...)

get_resid_variance(x, ...)

get_fixef_variance(x, ...)

get_ranef_variance(x, ...)

get_dist_variance(x, ...)

get_disp_variance(x, ...)
}
\arguments{
\item{x}{A mixed effects model of class \code{merMod}, \code{glmmTMB} or
\code{stanreg}.}

\item{component}{Character value, indicating the variance component that should
be returned. May be one of \code{"all"}, \code{"fixef"}, \code{"ranef"},
\code{"resid"}, \code{"dist"} or \code{"disp"}. By default, all variance
components are returned. The distribution-specific (\code{"dist"}) and
residual (\code{"resid"} variance are the most computational intensive
components, and hence may take a few seconds to calculate.}

\item{...}{Currently not used.}
}
\value{
A list with following elements:
   \itemize{
     \item \code{var.fixef}, variance attributable to the fixed effects
     \item \code{var.ranef}, (mean) variance of random effects
     \item \code{var.resid}, residual variance (sum of dispersion and distribution)
     \item \code{var.dist}, distribution-specific variance
     \item \code{var.disp}, variance due to additive dispersion
   }
}
\description{
This function extracts the different variance components of a
  mixed model and returns the result as list. Functions like
  \code{get_resid_variance(x)} or \code{get_fixef_variance(x)} are shortcuts
  for \code{get_variances(x, component = "resid")} etc.
}
\details{
This function returns different variance components from mixed models,
  which are needed, for instance, to calculate r-squared measures or the
  intraclass-correlation coefficient (ICC). The distributional variance (or
  observation-level variance) is based on lognormal approximation,
  \code{log(1+var(x)/mu^2)} (see \cite{Nakagawa et al. 2017}. The random effect
  variances are actually \emph{mean} random effect variances, thus they
  reflect the "average" random effects variance for mixed models with random
  slopes or nested random effects (see \cite{Johnson et al. 2014}). Details
  about the calculation of the different variance components can be found
  in \cite{Johnson et al. 2014} and \cite{Nakagawa et al. 2017}.
}
\examples{
\dontrun{
library(lme4)
data(sleepstudy)
m <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)

get_variances(m)
get_fixef_variance(m)
get_resid_variance(m)}

}
\references{
\itemize{
 \item Johnson, P. C. D. (2014). Extension of Nakagawa & Schielzeth’s R2 GLMM to random slopes models. Methods in Ecology and Evolution, 5(9), 944–946. \doi{10.1111/2041-210X.12225}
 \item Nakagawa, S., Johnson, P. C. D., & Schielzeth, H. (2017). The coefficient of determination R2 and intra-class correlation coefficient from generalized linear mixed-effects models revisited and expanded. Journal of The Royal Society Interface, 14(134), 20170213. \doi{10.1098/rsif.2017.0213}
 }
}
