% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\docType{class}
\name{INSiGHT-class}
\alias{INSiGHT-class}
\title{An S4 class to store key information.}
\description{
An S4 class to store key information.
}
\section{Slots}{

\describe{
\item{\code{geneExpr}}{A list containing gene expression matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each gene expression matrix is of dimension \eqn{G \times N}, where \eqn{G} rows represent \eqn{G} genes and
\eqn{N} columns represent \eqn{N} spots across this sample.}

\item{\code{spaCoord}}{A list containing spatial coordinate matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each spatial coordinate matrix is of dimension \eqn{N \times 2},
where \eqn{N} rows represent \eqn{N} spots across this sample,
and 2 columns represents x axis and y axis.}

\item{\code{covariateMat}}{A list containing covariate matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each covariate matrix is of dimension \eqn{N \times p},
where \eqn{N} rows represent \eqn{N} spots across this sample,
and \eqn{p} columns represents \eqn{p} covariates.}

\item{\code{weightMat}}{A list containing weight matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each covariate matrix is of dimension \eqn{N \times N},
where \eqn{N} is the number of spots across this sample.}

\item{\code{samples}}{A character string vector containing the sample IDs of all the samples to analyze.
The vector is of length \eqn{M}, where \eqn{M} is the number of samples.}

\item{\code{genes}}{A character string vector containing the gene IDs shared acorss all the samples to analyze.
The vector is of length \eqn{G}, where \eqn{G} is the number of genes to analyze.}

\item{\code{barcodes}}{A list containing barcode (spot IDs) vectors of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each barcode vector contains all the barcodes (spot IDs) of a sample.}

\item{\code{bandwidths}}{A numeric vector containing bandwidth hyper-parameters for the kernel functions
to contruct spatial similarity matrices.}

\item{\code{W.info}}{A list containing the information of the weighted spatial similarity matrices,
including W matrices, the corresponding eigen-values, and etc.}

\item{\code{rankScores}}{A list containing rank score matrices of all the samples to analyze.
The list is of length \eqn{M}, where \eqn{M} is the number of samples.
Each rank score matrix is of dimension \eqn{G \times N}, where \eqn{G} rows represent \eqn{G} genes and
\eqn{N} columns represent \eqn{N} spots across this sample.}

\item{\code{testResults}}{A data frame containing the test results of all the genes analyzed.}

\item{\code{features}}{Contains other features}
}}

