% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{insight-options}
\alias{insight-options}
\title{Global options from the insight package}
\description{
Global options from the insight package
}
\section{Global options to set defaults for function arguments}{

\itemize{
\item \verb{options(easystats_display_format = <value>)} will set the default format
for the \code{display()} methods. Can be one of \code{"markdown"}, \code{"html"}, or
\code{"tt"}. See \code{\link[=display]{display()}} for details.
\item \verb{options(easystats_table_width = <value>)} sets the default width
for tables printed to the console. Tables exceeding this width are
wrapped into multiple parts to ensure clean formatting and avoid
excessively long lines.
\item \verb{options(easystats_ci_separator = <value>)} defines the separator string
used between the lower and upper CI limits. Defaults to \code{", "}, resulting
in an output like \verb{[0.3, 0.8]}.
\item \verb{options(easystats_ci_brackets = <value>)} determines whether brackets or
parentheses are used to enclose formatted CIs. Can be \code{FALSE} to omit them,
\code{TRUE} to include them (defaults to \code{[} and \verb{]} for text output), or a
character vector of length 2 to define custom opening and closing symbols
(e.g., \code{c("(", ")")}).
\item \verb{options(insight_use_symbols = <value>)} accepts \code{TRUE} or \code{FALSE}. If
\code{TRUE}, statistical symbols (e.g., eta-squared, omega) are rendered using
their corresponding Unicode characters.
\item \verb{options(easystats_errors = <value>)} accepts \code{TRUE} or \code{FALSE}. If
\code{TRUE}, functions that typically suppress errors (via \code{tryCatch()})
will instead halt execution and throw an error when an issue occurs.
}
}

