% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_object.R
\name{is_empty_object}
\alias{is_empty_object}
\title{Check if object is empty}
\usage{
is_empty_object(x)
}
\arguments{
\item{x}{A list, a vector, or a dataframe.}
}
\value{
A logical indicating whether the entered object is empty.
}
\description{
Check if object is empty
}
\examples{
is_empty_object(c(1, 2, 3, NA))
is_empty_object(list(NULL, c(NA, NA)))
is_empty_object(list(NULL, NA))
}
