% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_function.R
\name{link_function}
\alias{link_function}
\alias{link_function.stanmvreg}
\alias{link_function.brmsfit}
\title{Get link-function from model object}
\usage{
link_function(x, ...)

\method{link_function}{stanmvreg}(x, mv_response = FALSE, ...)

\method{link_function}{brmsfit}(x, mv_response = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{mv_response}{Logical, if \code{TRUE} and model is a multivariate response
model from a \code{brmsfit} object or of class \code{stanmvreg}, then a
list of values (one for each regression) is returned.}
}
\value{
A function, describing the link-function from a model-object.
}
\description{
Returns the link-function from a model object.
}
\examples{
# example from ?stats::glm
counts <- c(18, 17, 15, 20, 10, 20, 25, 13, 12)
outcome <- gl(3, 1, 9)
treatment <- gl(3, 3)
m <- glm(counts ~ outcome + treatment, family = poisson())

link_function(m)(.3)
# same as
log(.3)
}
