% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log-Likelihood}
\usage{
loglikelihood(x, estimator = "OLS", summarize = TRUE, ...)
}
\arguments{
\item{x}{A model.}

\item{estimator}{Corresponds to the different estimators for the standard deviation of the errors. If \code{estimator="OLS"} (default), it returns the unbiased OLS estimator (scaling by the degrees of freedom, i.e., \code{n−k ()}). If \code{estimator="ML"}, the scaling is done by n (the biased ML estimator), which is then equivalent to using \code{stats::logLik()}. In moderately large samples, the differences should be negligible, but it is possible that OLS would perform slightly better in small samples with Gaussian errors.}

\item{summarize}{Currently not used.}

\item{...}{Not used.}
}
\description{
A robust function to compute the log-likelihood of a model.
}
\examples{
x <- lm(Sepal.Length ~ Petal.Width + Species, data=iris)

loglikelihood(x, estimator = "OLS")
loglikelihood(x, estimator = "ML")  # Equivalent to stats::logLik(x)
}
