% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\alias{model_data.merMod}
\title{Get name of a model's response variable}
\usage{
model_data(x, ...)

\method{model_data}{merMod}(x, parms = c("fixed", "random", "all"), ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{parms}{Should model data for fixed effects, random effects
or both be returned? Only applies to mixed models.}
}
\value{
## TODO docs
}
\description{
Returns the name(s) of the response variable(s) from a model object.
}
\examples{
data(cbpp, package = "lme4")
cbpp$trials <- cbpp$size - cbpp$incidence
m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
head(model_data(m))

}
