% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predictors.R
\name{model_predictors}
\alias{model_predictors}
\alias{model_predictors.default}
\alias{model_predictors.glmmTMB}
\alias{model_predictors.MixMod}
\title{Get term names of model predictors}
\usage{
model_predictors(x, ...)

\method{model_predictors}{default}(x, effects = c("fixed", "random",
  "all"), ...)

\method{model_predictors}{glmmTMB}(x, effects = c("fixed", "random",
  "all"), zi = FALSE, disp = FALSE, ...)

\method{model_predictors}{MixMod}(x, effects = c("fixed", "random",
  "all"), zi = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{effects}{Should predictor variables for fixed effects, random effects
or both be returned? Only applies to mixed models.}

\item{zi}{Logical, if \code{TRUE} and model has a zero-inflation-formula,
the variable(s) used in this formula are also returned.}

\item{disp}{Logical, if \code{TRUE} and model is of class \code{glmmTMB} and
has a dispersion-formula, the variable(s) used in the dispersion-formula
are also returned.}
}
\value{
The name(s) of the predictor variables from \code{x} as character vector.
}
\description{
to do...
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
model_predictors(m)

}
