% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_response.R
\name{model_response}
\alias{model_response}
\alias{model_response.default}
\title{Get name of a model's response variable}
\usage{
model_response(x, ...)

\method{model_response}{default}(x, combine = TRUE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{combine}{Logical, if \code{TRUE} and the response is a matrix-column,
the name of the response matches the notation in formula, and would for
instance also contain patterns like \code{"cbind(...)"}. Else, the original
variable names from the matrix-column are returned. See 'Examples'.}
}
\value{
The name(s) of the response variable(s) from \code{x} as character vector.
}
\description{
Returns the name(s) of the response variable(s) from a model object.
}
\examples{
library(lme4)
data(cbpp)
cbpp$trials <- cbpp$size - cbpp$incidence
m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)

model_response(m, combine = TRUE)
model_response(m, combine = FALSE)

}
