% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_terms.R
\name{model_terms}
\alias{model_terms}
\title{Get name of all model terms}
\usage{
model_terms(x, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}
}
\value{
Returns the name(s) of all model terms from \code{x} as character vector.
}
\description{
Returns the name(s) of all model terms, including response value
  and random effects.
}
\examples{
library(lme4)
data(cbpp)
cbpp$trials <- cbpp$size - cbpp$incidence

m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
model_terms(m)

m <- glmer(
  cbind(incidence, size - incidence) ~ period + (1 | herd),
  data = cbpp,
  family = binomial
)
model_terms(m)

}
