% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{nhanes_sample}
\alias{nhanes_sample}
\title{Sample dataset from the National Health and Nutrition Examination Survey}
\description{
Selected variables from the National Health and Nutrition Examination
             Survey that are used in the example from Lumley (2010), Appendix E.
}
\references{
Lumley T (2010). Complex Surveys: a guide to analysis using R. Wiley
}
\keyword{data}
