% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{origin_function}
\alias{origin_function}
\title{Find from which package a function comes from.}
\usage{
origin_function(fn)
}
\arguments{
\item{fn}{Function whose origin we want to find.}
}
\value{
The name of the package
}
\description{
This is useful when there are namespace conflicts and we want to know from
which package a function comes from.
}
