% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_color.R
\name{print_color}
\alias{print_color}
\alias{print_colour}
\alias{color_text}
\alias{colour_text}
\title{Coloured console output}
\usage{
print_color(text, color)

print_colour(text, colour)

color_text(text, color)

colour_text(text, colour)
}
\arguments{
\item{text}{The text to print.}

\item{color, colour}{Character vector, indicating the colour for printing.
May be one of \code{"red"}, \code{"yellow"}, \code{"green"}, \code{"blue"},
\code{"violet"}, \code{"cyan"} or \code{"grey"}. Formatting is also possible
with \code{"bold"} or \code{"italic"}.}
}
\value{
Nothing.
}
\description{
Convenient function that allows coloured output in the console.
 Mainly implemented to reduce package dependencies.
}
\details{
This function prints \code{text} directly to the console using
  \code{cat()}, so no string is returned. \code{color_text()}, however,
  returns only the formatted string, without using \code{cat()}.
}
\examples{
print_color("I'm blue dabedi dabedei", "blue")
}
