% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multiple_parameters.R
\name{print_multiple_parameters}
\alias{print_multiple_parameters}
\title{Get clean names of model paramters}
\usage{
print_multiple_parameters(x, ..., split_by = c("Effects", "Component",
  "Group", "Response"))
}
\arguments{
\item{x}{A fitted model.}

\item{...}{One or more objects (data frames), which contain information about
the model parameters and related statistics (like confidence intervals, HDI,
ROPE, ...).}

\item{split_by}{If \code{NULL}, the merged data frame is returned. Else, the
data frame is split into a list, split by the values from those columns defined
in \code{split_by}. \code{split_by} should be a character vector with one or
more of the following elements: \code{"Effects"}, \code{"Component"},
\code{"Response"} and \code{"Group"}.}
}
\value{
A data frame or a list.
}
\description{
This function
}
\examples{
library(bayestestR)
model <- download_model("brms_zi_2")
x <- hdi(model, effects = "all", component = "all")
x
print_multiple_parameters(model, x)

}
