% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\name{QualityControl}
\alias{QualityControl}
\title{Do quality control for a INSiGHT object
Remove spots with low total expression count, low-expressed genes and mt genes.}
\usage{
QualityControl(object, spot.threshold = 10, gene.threshold = 0.05)
}
\arguments{
\item{object}{INSiGHT object}

\item{spot.threshold}{(default 10) filter out the spots whose total expression count
lower than the threshold.}

\item{gene.threshold}{(default 0.05) filter out low-expressed genes
who has a percentage of non-zero expression spots lower than the threshold.}
}
\value{
A INSiGHT object.
}
\description{
Do quality control for a INSiGHT object
Remove spots with low total expression count, low-expressed genes and mt genes.
}
