% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_terms.R
\name{re_terms}
\alias{re_terms}
\title{Get names of random effect terms}
\usage{
re_terms(x, ...)
}
\arguments{
\item{x}{A fitted mixed model.}

\item{...}{Currently not used.}
}
\value{
The name(s) of the random effects as character vector.
}
\description{
to do...
}
\examples{
## TODO

}
