% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_ci.R
\name{reshape_ci}
\alias{reshape_ci}
\title{Reshape CI between wide/long formats}
\usage{
reshape_ci(x)
}
\arguments{
\item{x}{A data frame containing columns named \code{CI_low} and \code{CI_high}.}
}
\description{
Reshape CI between wide/long formats.
}
\examples{
x <- data.frame(
  Parameter = c("Term 1", "Term 2", "Term 1", "Term 2"),
  CI = c(.8, .8, .9, .9),
  CI_low = c(.2, .3, .1, .15),
  CI_high = c(.5, .6, .8, .85),
  stringsAsFactors = FALSE
)
reshape_ci(x)
reshape_ci(reshape_ci(x))
}
