% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_data.R
\name{response_data}
\alias{response_data}
\alias{response_data.default}
\title{Get the values from a model's response variable}
\usage{
response_data(x, ...)

\method{response_data}{default}(x, resp = NULL, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{resp}{Optional names of response variables for which to extract values.}
}
\value{
The values of the reponse variable, as vector, or a data frame if
  \code{x} has more than one defined response variable.
}
\description{
Returns the values the response variable(s) from a model object.
   If the model is a multivariate response model, a data frame with values
   from all response variables is returned.
}
\examples{
library(lme4)
data(cbpp)
data(mtcars)
cbpp$trials <- cbpp$size - cbpp$incidence

m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
response_data(m)
response_data(m, resp = "incidence")

m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
response_data(m)

}
