% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\name{testINSiGHT}
\alias{testINSiGHT}
\title{Run INSiGHT for multi-sample integative detection of SVGs}
\usage{
testINSiGHT(
  object,
  QuadForm.method = c("davies", "liu"),
  p.adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")
)
}
\arguments{
\item{object}{INSiGHT object.}

\item{QuadForm.method}{A character string specifying the method
to compute p-values for a quadratic form.}

\item{p.adjust.method}{A character string specifying the correction methods
for FDR control, including "holm", "hochberg", "hommel", "bonferroni", "BH",
"BY", "fdr" and "none". See \code{p.adjust.methods} for more details.}
}
\value{
A INSiGHT object containing test results.
}
\description{
Run INSiGHT hypothesis tests to detect spatially variable genes (SVGs)
for multiple samples.
}
\keyword{internal}
