% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{tidy}
\alias{tidy}
\title{Tidy methods for easystats-objects}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{A data frame, as returned by functions from easystats-packages like \code{\link[parameters:model_parameters]{model_parameters()}} or \code{\link[effectsize:effectsize]{effectsize()}}.}

\item{...}{Currently not used.}
}
\value{
\code{x}, with "standardized" column names (see \code{\link{standardize_names}}).
}
\description{
Tidy methods for easystats-objects. \code{tidy()} usually returns
  the same data frame that was used as input, however, with broom-alike
  column names.
}
\note{
Due to possible namespace conflicts with other packages that define
  a generic \code{tidy()}-method, the lifecycle of this function is still
  experimental, and there might be a chance that it will become defunct and
  removed in a future update. If possible, the preferred and stable alternative
  to \code{tidy()} for \emph{easystats}-objects is \code{\link{standardize_names}}.
}
\examples{
if (require("parameters")) {
  model <- lm(mpg ~ wt + cyl, data = mtcars)
  mp <- model_parameters(model)

  as.data.frame(mp)
  tidy(mp)
}
}
