% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawizard.R
\name{to_numeric}
\alias{to_numeric}
\title{Convert to Numeric (if possible)}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{A vector to be converted.}
}
\value{
Numeric vector (if possible)
}
\description{
Tries to convert vector to numeric if possible (if no warnings or errors).
Otherwise, leaves it as is.
}
\examples{
to_numeric(c("1", "2"))
to_numeric(c("1", "2", "A"))
}
