% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_table.R
\name{to_table}
\alias{to_table}
\alias{table_to_markdown}
\title{Generic export of data frames into tables}
\usage{
to_table(x, ...)

table_to_markdown(x, ...)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Arguments passed to other methods.}
}
\value{
A data frame.
}
\description{
This is a generic function to export data frames into various
table formats (like plain text, markdown, ...). See the documentation
for the specific objects' classes.
}
