% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_parameters.R
\name{clean_parameters}
\alias{clean_parameters}
\title{Get clean names of model paramters}
\usage{
clean_parameters(x, ...)
}
\arguments{
\item{x}{A fitted model.}
}
\value{
A data frame with "cleaned" parameter names and information on
  effects, component and group where parameters belong to. To be consistent
  across different models, the returned data frame always has the four
  columns \code{parameter}, \code{effect}, \code{component} and \code{group},
  even for models where some of these columns are not applicable.
}
\description{
This function "cleans" names of model parameters by removing
patterns like \code{"r_"} or \code{"b[]"} (mostly applicable to Stan models)
and adding columns with information to which group or component parameters
belong (i.e. fixed or random, count or zero-inflated...)
}
\details{
The \code{effects} column indicate if a parameter is a \emph{fixed}
or \emph{random} effect. The \code{component} can either be \emph{conditional}
or \emph{zero_inflated}. For models with random effects, the \code{group}
column indicates the grouping factor of the random effects.
}
\examples{
model <- download_model("brms_zi_2")
clean_parameters(model)
}
