% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSiGHT.R
\name{constructGaussianW}
\alias{constructGaussianW}
\title{Construct a spatial similarity matrix using Gaussian kernel for a INSiGHT object}
\usage{
constructGaussianW(
  object,
  phi,
  NN4Kernel = FALSE,
  k.NN4Kernel = 100,
  eigen.less = FALSE,
  eigen.prop = 0.3,
  weighted = FALSE
)
}
\arguments{
\item{object}{INSiGHT object.}

\item{phi}{Bandwidth hyper-parameter for the Gaussian kernel.}

\item{NN4Kernel}{(default FALSE) logical; If true, only the pair-wise distances
of k-nearest-neighbor will be computed for each spot, resulting in a sparse matrix.
If false, all the pair-wise distances will be computed.}

\item{k.NN4Kernel}{The maximum number of nearest neighbors to compute
for the construction of Gaussian kernel using a kd-tree. Works when \code{NN4Kernel == TRUE}.}

\item{eigen.less}{(default FALSE) logical; If true, only the top eigen-values
will be computed with \code{eigs_sym} function to save computing time. Works when \code{NN4Kernel == TRUE}.
If false, all the eigen-values will be computed, which may be time-consuming
when the number of spots is large.}

\item{eigen.prop}{A number between \eqn{[0,1]}
specifying the proportion of top eigen-values to compute.
Works when \code{eigen.less == TRUE}.}

\item{weighted}{(default FALSE)
If true, then the each spatial similarity matrix will be weighted
by its corresponding weight matrix.}
}
\value{
A INSiGHT object.
}
\description{
Construct a spatial similarity matrix for a INSiGHT object,
using a distance-based Gaussian kernel function
(i.e. \eqn{K(s_i, s_j) = \exp(-\frac{\lVert s_i - s_j \rVert^2}{2 \phi^2})}),
where \eqn{\phi} is the bandwidth parameter.
}
