% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{data_match}
\alias{data_match}
\title{Find rows of a dataframe that are matching a specific subset}
\usage{
data_match(x, to)
}
\arguments{
\item{x}{A dataframe.}

\item{to}{The dataframe of which to meet the characteristics.}
}
\description{
Find row indices of a dataframe that are matching a specific configuration.
}
\examples{
matching_rows <- data_match(mtcars, data.frame(vs = 0, am = 1))
mtcars[matching_rows, ]

matching_rows <- data_match(mtcars, data.frame(vs = 0, am = c(0, 1)))
mtcars[matching_rows, ]

}
