% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_long.R
\name{data_to_long}
\alias{data_to_long}
\title{Reshape (pivot) data from wide to long}
\usage{
data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL
)
}
\arguments{
\item{data}{A data frame to pivot.}

\item{cols}{A vector of column names or indices to pivot into longer format.}

\item{colnames_to}{The name of the new column that will contain the column names.}

\item{values_to}{The name of the new column that will contain the values of the pivotted variables.}

\item{rows_to}{The name of the column that will contain the row-number from the original data. If \code{NULL}, will be removed.}
}
\description{
This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of \code{tidyr::pivot_longer()}.
}
\examples{
data <- data.frame(replicate(5, rnorm(10)))

# Default behaviour (equivalent to tidyr::pivot_longer(data, cols = 1:5))
data_to_long(data)

# Customizing the names
data_to_long(data,
             cols = c(1, 2),
             colnames_to = "Column",
             values_to = "Numbers",
             rows_to = "Row")


}
