% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\alias{print_md}
\alias{print_html}
\alias{display.data.frame}
\alias{print_md.data.frame}
\title{Generic export of data frames into formatted tables}
\usage{
display(object, ...)

print_md(x, ...)

print_html(x, ...)

\method{display}{data.frame}(object, ...)

\method{print_md}{data.frame}(x, ...)

\method{print_md}{data.frame}(x, ...)
}
\arguments{
\item{object, x}{A data frame.}

\item{...}{Arguments passed to other methods.}
}
\value{
A data frame.
}
\description{
\code{display()} is a generic function to export data frames
into various table formats (like plain text, markdown, ...). \code{print_md()}
usually is a convenient wrapper for \code{display(format = "markdown")}.
Similar, \code{print_html()} is a shortcut for \code{display(format = "html")}.
See the documentation for the specific objects' classes.
}
\examples{
display(iris)


}
