% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\alias{print_md}
\title{Generic export of data frames into formatted tables}
\usage{
display(object, ...)

print_md(x, ...)
}
\arguments{
\item{object, x}{A data frame.}

\item{...}{Arguments passed to other methods.}
}
\value{
A data frame.
}
\description{
\code{display()} is a generic function to export data frames
into various table formats (like plain text, markdown, ...). \code{print_md()}
usually is a convenient wrapper for \code{display(format = "markdown")}. See
the documentation for the specific objects' classes.
}
