% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model.R
\name{download_model}
\alias{download_model}
\title{Download circus models}
\usage{
download_model(name)
}
\arguments{
\item{name}{Model name.}
}
\value{
A model from the \emph{circus}-repository.
}
\description{
Downloads pre-compiled models from the \emph{circus}-repository.
  The \emph{circus}-repository contains a variety of fitted models to help
  the systematic testing of other packages
}
\references{
\url{https://easystats.github.io/circus/}
}
