% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model.R
\name{download_model}
\alias{download_model}
\title{Download circus models}
\usage{
download_model(name, url = NULL)
}
\arguments{
\item{name}{Model name.}

\item{url}{String with the URL from where to download the model data.
Optional, and should only be used in case the repository-URL is
changing. By default, models are downloaded from
\verb{https://raw.github.com/easystats/circus/master/data/}.}
}
\value{
A model from the \emph{circus}-repository.
}
\description{
Downloads pre-compiled models from the \emph{circus}-repository.
The \emph{circus}-repository contains a variety of fitted models to help
the systematic testing of other packages
}
\details{
The code that generated the model is available at the
\url{https://easystats.github.io/circus/reference/index.html}.
}
\references{
\url{https://easystats.github.io/circus/}
}
