% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_auxiliary.R
\name{find_auxiliary}
\alias{find_auxiliary}
\alias{find_auxiliary.default}
\title{Find auxiliary (distributional) parameters from models}
\usage{
find_auxiliary(x, ...)

\method{find_auxiliary}{default}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{A model of class \code{brmsfit}.}

\item{...}{Currently not used.}

\item{verbose}{Toggle warnings.}
}
\value{
The names of all available auxiliary parameters used in the model, or
\code{NULL} if no auxiliary parameters were found.
}
\description{
Returns the names of all auxiliary / distributional parameters
from brms-models, like dispersion, sigma, kappa, phi, or beta... For
univariate non-linear models (\code{nl = TRUE}), the non-linear parameters of \code{mu}
are not auxiliary parameters, and are not returned.
}
