% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_formula.R
\name{find_formula}
\alias{find_formula}
\title{Find model formula}
\usage{
find_formula(x, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}
}
\value{
A formula that describes the model, or a list of formulas (e.g., for
   multivariate response models, or models with zero-inflation component).
}
\description{
Get model formula
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_formula(m)

}
