% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_formula.R
\name{find_formula}
\alias{find_formula}
\title{Find model formula}
\usage{
find_formula(x, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}
}
\value{
A list of formulas that describe the model. For simple models,
   only one list-element, \code{conditional}, is returned. For more complex
   models, the returned list may have following elements:
   \itemize{
     \item \code{conditional}, the "fixed effects" part from the model
     \item \code{random}, the "random effects" part from the model
     \item \code{zero_inflated}, the "fixed effects" part from the zero-inflation component of the model
     \item \code{zero_inflated_random}, the "random effects" part from the zero-inflation component of the model
     \item \code{dispersion}, the dispersion formula
   }
}
\description{
Get model formula
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_formula(m)
}
