% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parameters_gam.R
\name{find_parameters.gamlss}
\alias{find_parameters.gamlss}
\title{Find names of model parameters from generalized additive models}
\usage{
\method{find_parameters}{gamlss}(x, flatten = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{flatten}{Logical, if \code{TRUE}, the values are returned
as character vector, not as list. Duplicated values are removed.}

\item{...}{Currently not used.}
}
\value{
A list of parameter names. The returned list may have following
  elements:
   \itemize{
     \item \code{conditional}, the "fixed effects" part from the model.
     \item \code{smooth_terms}, the smooth parameters.
   }
}
\description{
Returns the names of model parameters, like they typically
    appear in the \code{summary()} output.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_parameters(m)
}
