% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_predictors.R
\name{find_predictors}
\alias{find_predictors}
\title{Find names of model predictors}
\usage{
find_predictors(x, effects = c("fixed", "random", "all"),
  component = c("all", "conditional", "zi", "zero_inflated",
  "dispersion"), flatten = FALSE)
}
\arguments{
\item{x}{A fitted model.}

\item{effects}{Should variables for fixed effects, random effects
or both be returned? Only applies to mixed models. May be abbreviated.}

\item{component}{Should all predictor variables, predictor variables for the
conditional model, the zero-inflated part of the model or the dispersion
model be returned? Only applies to models with zero-inflated and/or
dispersion formula. May be abbreviated.}

\item{flatten}{Logical, if \code{TRUE}, the name of model terms are returned
as a single character, not as list.}
}
\value{
A list of character vectors that represent the name(s) of the
   predictor variables. Depending on the combination of the arguments
   \code{effects} and \code{component}, the returned list has following
   elements:
   \itemize{
     \item \code{conditional}, the "fixed effects" terms from the model
     \item \code{random}, the "random effects" terms from the model
     \item \code{zero_inflated}, the "fixed effects" terms from the zero-inflation component of the model
     \item \code{zero_inflated_random}, the "random effects" terms from the zero-inflation component of the model
     \item \code{dispersion}, the dispersion terms
   }
}
\description{
to do...
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_predictors(m)

}
