% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_terms.R
\name{find_terms}
\alias{find_terms}
\title{Find name of all model terms}
\usage{
find_terms(x, effects = c("all", "fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated",
  "dispersion"), flatten = FALSE)
}
\arguments{
\item{x}{A fitted model.}

\item{effects}{Should variables for fixed effects, random effects
or both be returned? Only applies to mixed models. May be abbreviated.}

\item{component}{Should all predictor variables, predictor variables for the
conditional model, the zero-inflated part of the model or the dispersion
model be returned? Only applies to models with zero-inflated and/or
dispersion formula. May be abbreviated.}

\item{flatten}{Logical, if \code{TRUE}, the name of model terms are returned
as a single character, not as list.}
}
\value{
A list with (depending on the model) following elements (character
   vectors):
   \itemize{
     \item \code{response}, the name of the response variable
     \item \code{conditional}, the name(s) of the predictor variables from the \emph{conditional} model (as opposed to the zero-inflated part of a model)
     \item \code{random}, the name of the random effects (grouping factors)
     \item \code{zero_inflated}, the name(s) of the predictor variables from the \emph{zero-inflated} part of the model
     \item \code{zero_inflated_random}, the name of the random effects (grouping factors)
     \item \code{disperion}, the name of the dispersion terms
   }
}
\description{
Returns a list with the name(s) of all model terms, including
  response value and random effects.
}
\examples{
library(lme4)
data(cbpp)
data(sleepstudy)
# some data preparation...
cbpp$trials <- cbpp$size - cbpp$incidence
sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
sleepstudy$mysubgrp <- NA
for (i in 1:5) {
  filter_group <- sleepstudy$mygrp == i
  sleepstudy$mysubgrp[filter_group] <-
    sample(1:30, size = sum(filter_group), replace = TRUE)
}

m1 <- glmer(
  cbind(incidence, size - incidence) ~ period + (1 | herd),
  data = cbpp,
  family = binomial
)
find_terms(m1)

m2 <- lmer(
  Reaction ~ Days + (1 | mygrp / mysubgrp) + (1 | Subject),
  data = sleepstudy
)
find_terms(m2)
find_terms(m2, flatten = TRUE)
}
