% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bf.R
\name{format_bf}
\alias{format_bf}
\title{Bayes Factor formatting}
\usage{
format_bf(
  bf,
  stars = FALSE,
  stars_only = FALSE,
  name = "BF",
  protect_ratio = FALSE,
  na_reference = NA
)
}
\arguments{
\item{bf}{Bayes Factor.}

\item{stars}{Add significance stars (e.g., p < .001***).}

\item{stars_only}{Return only significance stars.}

\item{name}{Name prefixing the text. Can be \code{NULL}.}

\item{protect_ratio}{Should values smaller than 1 be represented as ratios?}

\item{na_reference}{How to format missing values (\code{NA}). By default, \code{NA} values are formatted as \code{"BF = 1"}.}
}
\value{
A formatted string.
}
\description{
Bayes Factor formatting
}
\examples{
format_bf(1.20)
format_bf(c(1.20, 1557, 3.5, 12), stars = TRUE)
format_bf(c(1.20, 1557, 3.5, 12), name = NULL)
format_bf(c(NA, 3.45, 345))
format_bf(c(NA, 3.45, 345), na_reference = 1)
}
