% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_message.R
\name{format_message}
\alias{format_message}
\title{Format messages and warnings}
\usage{
format_message(string, line_length = options()$width)
}
\arguments{
\item{string}{A string.}

\item{line_length}{Numeric, the maximum length of a line.}
}
\value{
A formatted string.
}
\description{
Inserts line breaks into a longer message or warning string.
  Line length is adjusted to maximum length of the console, if the width
  can be accessed. By default, new lines are indented by two whitespace.
}
\examples{
format_message("Much too long string for just one line, I guess!", 15)
}
