% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{get_data}
\alias{get_data}
\alias{get_data.merMod}
\title{Get the data that was used to fit the model}
\usage{
get_data(x, ...)

\method{get_data}{merMod}(x, effects = c("fixed", "random", "all"), ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{effects}{Should model data for fixed effects, random effects
or both be returned? Only applies to mixed models.}
}
\value{
The data that was used to fit the model.
}
\description{
This functions tries to get the data that was used to fit the
  model and returns it as data frame.
}
\note{
Unlike \code{model.frame()}, which may contain transformed variables
  (e.g. if \code{poly()} or \code{scale()} was used inside the formula to
  specify the model), \code{get_data()} aims at returning the "original",
  untransformed data.
}
\examples{
data(cbpp, package = "lme4")
cbpp$trials <- cbpp$size - cbpp$incidence
m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
head(get_data(m))

}
