% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deviance.R
\name{get_deviance}
\alias{get_deviance}
\alias{get_deviance.default}
\title{Model Deviance}
\usage{
get_deviance(x, ...)

\method{get_deviance}{default}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Not used.}

\item{verbose}{Toggle warnings and messages.}
}
\value{
The model deviance.
}
\description{
Returns model deviance (see \code{stats::deviance()}).
}
\examples{
data(mtcars)
x <- lm(mpg ~ cyl, data = mtcars)
get_deviance(x)
}
