% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df.R
\name{get_df}
\alias{get_df}
\alias{get_df.default}
\title{Extract degrees of freedom}
\usage{
get_df(x, ...)

\method{get_df}{default}(x, method = "residual", verbose = TRUE, ...)
}
\arguments{
\item{x}{A statistical model.}

\item{...}{Currently not used.}

\item{method}{Can be \code{"residual"} or \code{"model"}. \code{"residual"}
tries to extract residual degrees of freedoms. If residual degrees of freedom
could not be extracted, returns \code{n-k} (number of observations minus
number of parameters). \code{"model"} returns model-based degrees of freedom,
i.e. the number of parameters (+1 for linear models for the sigma-parameter).}

\item{verbose}{Toggle warnings.}
}
\description{
Estimate or extract residual or model-based degrees of freedom from regression models.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
get_df(model)

model <- glm(vs ~ mpg * cyl, data = mtcars, family = "binomial")
get_df(model)
}
