% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_mixed.R
\name{get_parameters.glmm}
\alias{get_parameters.glmm}
\title{Get model parameters from mixed models}
\usage{
\method{get_parameters}{glmm}(x, effects = c("all", "fixed", "random"), ...)
}
\arguments{
\item{x}{A fitted model.}

\item{effects}{Should parameters for fixed effects, random effects
or both be returned? Only applies to mixed models. May be abbreviated.}

\item{...}{Currently not used.}

\item{verbose}{Toggle messages and warnings.}
}
\value{
If \code{effects = "fixed"}, a data frame with two columns: the
  parameter names and the related point estimates. If \code{effects = "random"},
  a list of data frames with the random effects (as returned by \code{ranef()}),
  unless the random effects have the same simplified structure as fixed effects
  (e.g. for models from \pkg{MCMCglmm}).
}
\description{
Returns the coefficients from a model.
}
\details{
In most cases when models either return different "effects" (fixed,
random) or "components" (conditional, zero-inflated, ...), the arguments
\code{effects} and \code{component} can be used.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_parameters(m)
}
