% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R, R/get_parameters_gam.R,
%   R/get_parameters_mfx.R, R/get_parameters_mixed.R, R/get_parameters_zi.R
\name{get_parameters}
\alias{get_parameters}
\alias{get_parameters.gam}
\alias{get_parameters.logitmfx}
\alias{get_parameters.coxme}
\alias{get_parameters.merMod}
\alias{get_parameters.mixed}
\alias{get_parameters.glmmTMB}
\alias{get_parameters.BBmm}
\alias{get_parameters.glimML}
\alias{get_parameters.zcpglm}
\title{Get model parameters}
\usage{
get_parameters(x, ...)

\method{get_parameters}{gam}(
  x,
  component = c("all", "conditional", "smooth_terms", "location"),
  ...
)

\method{get_parameters}{logitmfx}(x, component = c("all", "conditional", "marginal"), ...)

\method{get_parameters}{coxme}(x, effects = c("fixed", "random"), ...)

\method{get_parameters}{merMod}(x, effects = c("fixed", "random"), ...)

\method{get_parameters}{mixed}(x, effects = c("fixed", "random"), ...)

\method{get_parameters}{glmmTMB}(
  x,
  effects = c("fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
  ...
)

\method{get_parameters}{BBmm}(x, effects = c("fixed", "random"), ...)

\method{get_parameters}{glimML}(x, effects = c("fixed", "random", "all"), ...)

\method{get_parameters}{zcpglm}(
  x,
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{component}{Which type of parameters to return, such as parameters for the
 conditional model, the zero-inflated part of the model, the dispersion
 term, the instrumental variables or marginal effects be returned? Applies
 to models with zero-inflated and/or dispersion formula, or to models with
 instrumental variables (so called fixed-effects regressions), or models
 with marginal effects from \pkg{mfx}. May be abbreviated. Note that the
\emph{conditional} component is also called \emph{count} or \emph{mean}
component, depending on the model. There are two convenient shortcuts:
If \code{component = "location"}, location parameters such as \code{conditional},
\code{zero_inflated}, \code{smooth_terms}, or \code{instruments} are returned.
For \code{component = "distributional"} (or \code{"auxiliary"}), components
like \code{sigma}, \code{dispersion}, \code{beta} or \code{precision} (and
other auxiliary parameters) are returned.}

\item{effects}{Should parameters for fixed effects, random effects
or both be returned? Only applies to mixed models. May be abbreviated.}

\item{verbose}{Toggle messages and warnings.}
}
\value{
\itemize{
  \item for non-Bayesian models, a data frame with two columns: the parameter names and the related point estimates.
  \item for Anova (\code{aov()}) with error term, a list of parameters for the conditional and the random effects parameters
}
}
\description{
Returns the coefficients (or posterior samples for Bayesian
   models) from a model. See the documentation for your object's class:
   \itemize{
     \item{\link[=get_parameters.BGGM]{Bayesian models} (\pkg{rstanarm}, \pkg{brms}, \pkg{MCMCglmm}, ...)}
     \item{\link[=get_parameters.emmGrid]{Estimated marginal means} (\pkg{emmeans})}
     \item{\link[=get_parameters.gamm]{Generalized additive models} (\pkg{mgcv}, \pkg{VGAM}, ...)}
     \item{\link[=get_parameters.betamfx]{Marginal effects models} (\pkg{mfx})}
     \item{\link[=get_parameters.glmm]{Mixed models} (\pkg{lme4}, \pkg{glmmTMB}, \pkg{GLMMadaptive}, ...)}
     \item{\link[=get_parameters.zeroinfl]{Zero-inflated and hurdle models} (\pkg{pscl}, ...)}
     \item{\link[=get_parameters.betareg]{Models with special components} (\pkg{betareg}, \pkg{MuMIn}, ...)}
   }
}
\details{
In most cases when models either return different "effects" (fixed,
random) or "components" (conditional, zero-inflated, ...), the arguments
\code{effects} and \code{component} can be used.
\cr \cr
\code{get_parameters()} is comparable to \code{coef()}, however, the coefficients
are returned as data frame (with columns for names and point estimates of
coefficients). For Bayesian models, the posterior samples of parameters are
returned.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_parameters(m)
}
