% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted.R
\name{get_predicted}
\alias{get_predicted}
\title{Predicted values}
\usage{
get_predicted(x, newdata = NULL, ...)
}
\arguments{
\item{x}{A model.}

\item{newdata}{An optional data frame in which to look for variables with
    which to predict.  If omitted, the fitted values are used.}

\item{...}{Not used.}
}
\value{
The fitted values (i.e. predictions for the response).
}
\description{
Returns values predicted by a model (i.e., fitted values).
}
\note{
Currently, this function just calls \code{stats::fitted()}, but will
be extended to other objects that don't work with \code{stats::fitted()} in
future updates.
}
\examples{
data(mtcars)
x <- lm(mpg ~ cyl + hp, data = mtcars)
get_predicted(x)
}
