% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted.R
\name{get_predicted}
\alias{get_predicted}
\title{Predicted values}
\usage{
get_predicted(x, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Passed down to \code{residuals()}, if possible.}
}
\description{
Returns values predicted by a model (i.e., fitted values).
}
\examples{
x <- lm(Sepal.Length ~ Petal.Width * Species, data=iris)
get_predicted(x)
}
