% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{Get the data from model predictors}
\usage{
get_predictors(x, verbose = TRUE)
}
\arguments{
\item{x}{A fitted model.}

\item{verbose}{Toggle messages and warnings.}
}
\value{
The data from all predictor variables, as data frame.
}
\description{
Returns the data from all predictor variables (fixed effects).
}
\examples{
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
head(get_predictors(m))
}
