% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random.R
\name{get_random}
\alias{get_random}
\title{Get the data from a model's random effects terms}
\usage{
get_random(x)
}
\arguments{
\item{x}{A fitted mixed model.}
}
\value{
The data of all random effects terms, as data frame.
}
\description{
Returns the data of all random effect terms.
}
\examples{
library(lme4)
data(sleepstudy)
# prepare some data...
sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
sleepstudy$mysubgrp <- NA
for (i in 1:5) {
  filter_group <- sleepstudy$mygrp == i
  sleepstudy$mysubgrp[filter_group] <-
  sample(1:30, size = sum(filter_group), replace = TRUE)
}

m <- lmer(
  Reaction ~ Days + (1 | mygrp / mysubgrp) + (1 | Subject),
  data = sleepstudy
)

get_random(m)

}
