% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Get the values from a model's response variable}
\usage{
get_response(x, resp = NULL)
}
\arguments{
\item{x}{A fitted model.}

\item{resp}{Optional names of response variables for which to extract values.}
}
\value{
The values of the reponse variable, as vector, or a data frame if
  \code{x} has more than one defined response variable.
}
\description{
Returns the values the response variable(s) from a model object.
   If the model is a multivariate response model, a data frame with values
   from all response variables is returned.
}
\examples{
library(lme4)
data(cbpp)
data(mtcars)
cbpp$trials <- cbpp$size - cbpp$incidence

m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
get_response(m)
get_response(m, resp = "incidence")

m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_response(m)

}
