% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Get residual standard deviation from models}
\usage{
get_sigma(x)
}
\arguments{
\item{x}{A model.}
}
\value{
The residual standard deviation (sigma)
}
\description{
Returns the residual standard deviation from classical
  and mixed models.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_sigma(m)
}
