% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varcov.R
\name{get_varcov}
\alias{get_varcov}
\alias{get_varcov.hurdle}
\alias{get_varcov.MixMod}
\alias{get_varcov.glmmTMB}
\alias{get_varcov.brmsfit}
\title{Get variance-covariance matrix from models}
\usage{
get_varcov(x, ...)

\method{get_varcov}{hurdle}(x, component = c("conditional",
  "zero_inflated", "zi", "all"), ...)

\method{get_varcov}{MixMod}(x, component = c("conditional",
  "zero_inflated", "zi", "all"), ...)

\method{get_varcov}{glmmTMB}(x, component = c("conditional",
  "zero_inflated", "zi", "all"), ...)

\method{get_varcov}{brmsfit}(x, component = c("conditional",
  "zero_inflated", "zi", "all"), ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Currently not used.}

\item{component}{Should the complete variance-covariance matrix of the model
be returned, or only for the count or zero-inlfated model? Applies to models
with zero-inflated component. \code{component} may be one of
\code{"conditional"}, \code{"zi"}, \code{"zero-inflated"} or \code{"all"}.
May be abbreviated.}
}
\value{
The variance-covariance matrix, as \code{matrix}-object.
}
\description{
Returns the variance-covariance, as retrieved by
  \code{stats::vcov()}, but works for more model objects that probably
  don't provide a \code{vcov()}-method.
}
\note{
\code{get_varcov()} tries to return the nearest positive definite matrix
  in case of a negative variance-covariance matrix.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_varcov(m)
}
