% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_inverse.R
\name{link_inverse}
\alias{link_inverse}
\alias{link_inverse.stanmvreg}
\alias{link_inverse.brmsfit}
\title{Get link-inverse function from model object}
\usage{
link_inverse(x, ...)

\method{link_inverse}{stanmvreg}(x, mv_response = FALSE, ...)

\method{link_inverse}{brmsfit}(x, mv_response = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{mv_response}{Logical, if \code{TRUE} and model is a multivariate response
model from a \code{brmsfit} object or of class \code{stanmvreg}, then a
list of values (one for each regression) is returned.}
}
\value{
A function, describing the inverse-link function from a model-object.
}
\description{
Returns the link-inverse function from a model object.
}
\examples{
# example from ?stats::glm
counts <- c(18, 17, 15, 20, 10, 20, 25, 13, 12)
outcome <- gl(3, 1, 9)
treatment <- gl(3, 3)
m <- glm(counts ~ outcome + treatment, family = poisson())

link_inverse(m)(.3)
# same as
exp(.3)

}
