% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{model_info}
\alias{model_info}
\alias{model_info.brmsfit}
\title{Access information from model objects}
\usage{
model_info(x, ...)

\method{model_info}{brmsfit}(x, mv_response = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{mv_response}{Logical, if \code{TRUE} and model is a multivariate response
model from a \code{brmsfit} object or of class \code{stanmvreg}, then a
list of values (one for each regression) is returned.}
}
\value{
A list with information about the model, like family, link-function
  etc. (see 'Details').
}
\description{
Retrieve information from model objects.
}
\details{
\code{model_info()} returns a list with information about the
  model for many different model objects. Following information
   is returned, where all values starting with \code{is_} are logicals.
   \itemize{
     \item \code{is_binomial}: family is binomial (but not negative binomial)
     \item \code{is_poisson}: family is poisson
     \item \code{is_negbin}: family is negative binomial
     \item \code{is_count}: model is a count model (i.e. family is either poisson or negative binomial)
     \item \code{is_beta}: family is beta
     \item \code{is_logit}: model has logit link
     \item \code{is_linear}: family is gaussian
     \item \code{is_ordinal}: family is ordinal or cumulative link
     \item \code{is_categorical}: family is categorical link
     \item \code{is_zeroinf}: model has zero-inflation component
     \item \code{is_multivariate}: model is a multivariate response model (currently only works for \emph{brmsfit} objects)
     \item \code{is_trial}: model response contains additional information about the trials
     \item \code{is_bayesian}: model is a Bayesian model
     \item \code{link_fun}: the link-function
     \item \code{family}: the family-object
     \item \code{nobs}: number of observations
     \item \code{model_terms}: a list with all model terms, including terms such as random effects or from zero-inflated model parts.
   }
}
\examples{
library(glmmTMB)
data("Salamanders")
m <- glmmTMB(
  count ~ spp + cover + mined + (1 | site),
  ziformula = ~ spp + mined,
  dispformula = ~DOY,
  data = Salamanders,
  family = nbinom2
)

model_info(m)
}
