% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_name.R
\name{model_name}
\alias{model_name}
\title{Name the model}
\usage{
model_name(x, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Passed down to \code{residuals()}, if possible.}
}
\value{
The value of the intercept.
}
\description{
Returns the value at the intercept (i.e., the intercept parameter), and \code{NA} if there isn't one.
}
\examples{
model_name(lm(Sepal.Length ~ Petal.Width, data = iris))

if (require("lme4")) {
  model_name(lme4::lmer(Sepal.Length ~ Sepal.Width + (1 | Species), data = iris))
}
if (require("rstanarm")) {
  model_name(rstanarm::stan_glm(Sepal.Length ~ Petal.Width,
  data = iris, refresh = 0, iter = 200))
}
# if (require("gamm4")) {
#   model_name(gamm4::gamm4(Sepal.Length ~ s(Petal.Width), data = iris))
#}
}
