% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_random.R
\name{model_random}
\alias{model_random}
\title{Get names of random effect terms}
\usage{
model_random(x, ...)
}
\arguments{
\item{x}{A fitted mixed model.}

\item{...}{Currently not used.}
}
\value{
The name(s) of the random effects as character vector.
}
\description{
to do...
}
\examples{
## TODO

}
