% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_tools.R
\name{print_color}
\alias{print_color}
\alias{print_colour}
\title{Coloured console output}
\usage{
print_color(color, text)

print_colour(colour, text)
}
\arguments{
\item{color, colour}{Character vector, indicating the colour for printing.
May be one of \code{"red"}, \code{"yellow"}, \code{"green"}, \code{"blue"},
\code{"violet"}, \code{"cyan"} or \code{"grey"}. Formatting is also possible
with \code{"bold"} or \code{"italic"}.}

\item{text}{The text to print.}
}
\value{
Nothing.
}
\description{
Convenient function that allows coloured output in the console.
 Mainly implemented to reduce package dependencies.
}
\details{
This function prints \code{text} directly to the console using
  \code{cat()}, so no string is returned.
}
\examples{
print_color("blue", "I'm blue dabedi dabedei")

}
