% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovFPC.R
\name{vcovFPC}
\alias{vcovFPC}
\alias{vcovFPC.lm}
\alias{vcovFPC.merMod}
\title{Obtained finite-population-adjusted standard errors}
\usage{
vcovFPC(object, ..., verbose = TRUE)

\method{vcovFPC}{lm}(object, popsize = NULL, ..., verbose = TRUE)

\method{vcovFPC}{merMod}(object, popsize2 = NULL, popsize1 = NULL, ..., verbose = TRUE)
}
\arguments{
\item{object}{a model inheriting from \code{lm} or \code{merMod} (2-level only).}

\item{verbose}{Whether to print messages about whether FPC is applied or not.}

\item{popsize, popsize1, popsize2}{The finite population size (for \code{merMod}
specified seperetly for level-1 and level-2). If \code{NULL}, an infinite
population is assumed.}
}
\value{
The variance-covariance matrix of the fixed effect estimates, as
returned by \code{vcov()}, but with FPC applied.
}
\description{
This function return the variance-covariance matrix as returned by \code{vcov()},
but with finite population correction (FPC) applied.
}
\details{
The FPC is defined as:
\cr\cr
\deqn{FPC = \frac{N - n}{N - 1}}{FPC = (N - n) / (N - 1)}
\cr\cr
FPC for multilevel models is based on the method described by Lai et al. (2018).
}
\references{
Lai, M. H., Kwok, O. M., Hsiao, Y. Y., & Cao, Q. (2018). Finite population correction for two-level hierarchical linear models. \emph{Psychological methods, 23}(1), 94.
}
