% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varcov_fpc.R
\name{vcovFPC}
\alias{vcovFPC}
\alias{vcovFPC.merMod}
\title{Obtain finite-population-adjusted variance-covariance matrix}
\usage{
vcovFPC(model, ...)

\method{vcovFPC}{merMod}(model, population_size = NULL, cluster_size = NULL, kr = FALSE, ...)
}
\arguments{
\item{model}{A model inheriting from \code{lm}, \code{glm}, \code{merMod} or \code{glmmTMB}
(2-level only).}

\item{...}{Not used.}

\item{population_size}{The finite population size (for mixed models: for
level-1).}

\item{cluster_size}{The finite population size for level-2 (cluster groups).}

\item{kr}{Logical, if \code{TRUE}, also applies Kenward-Roger adjustment to the
returned variance-covariance matrix.}
}
\value{
The variance-covariance matrix of the fixed effect estimates, as
returned by \code{vcov()}, but with FPC applied.
}
\description{
This function returns the variance-covariance matrix as returned by \code{vcov()},
but with finite population correction (FPC) applied.
}
\details{
The FPC is defined as:

\deqn{FPC = \frac{N - n}{N - 1}}{FPC = (N - n) / (N - 1)}

FPC for multilevel models is based on the method described by Lai et al. (2018).
}
\references{
Lai, M. H., Kwok, O. M., Hsiao, Y. Y., & Cao, Q. (2018). Finite
population correction for two-level hierarchical linear models.
\emph{Psychological methods, 23}(1), 94.
}
